Removed obsolete ValidationWrapper component
