--- src/components/ValidationWrapper/index.ts
@@ -1,2 +1,0 @@
-export * from './ValidationWrapper';
-export * from './ValidationWrapper.types';
